fastapi_app/pyproject.toml
[project]
name = "fastapi-app"
version = "0.1.0"
description = "FastAPI backend base for the full-stack app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.135.2",
    "pypdf>=5.0.0",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.43",
    "groq>=0.15.0",
    "pgvector>=0.3.0",
    "PyJWT>=2.9.0",
    "slowapi>=0.1.9",
    "sentence-transformers>=3.4.0",
    "uvicorn[standard]>=0.35.0",
    "apscheduler>=3.10.4",
    "httpx>=0.28.1",
    "easyocr>=1.7.0",
    "opencv-python-headless>=4.9.0",
    "email-validator>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.12.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["F"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
